+++ src/system/libretro/libretro-common/include/retro_inline.h
