+++ src/system/libretro/libretro_core_options_intl.h
@@ -7,7 +7,7 @@
 #pragma warning(disable:4566)
 #endif
 
-#include "libretro.h"
+#include "libretro-common/include/libretro.h"
 
 /*
  ********************************
